server/requirements.txt
Flask
flask-cors
flask-socketio
gunicorn
eventlet
python-dotenv
deep-translator
textblob